requirements.txt
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank-bm25==0.2.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
